API_files/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
elastic-transport==8.17.0
elasticsearch==8.17.1
fastapi==0.115.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai==1.63.2
pydantic==2.10.6
pydantic_core==2.27.2
pyodbc==5.2.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.2
markdown==3.7.0
gunicorn==23.0.0